src/api/requirements.txt
fastapi
uvicorn[standard]
pillow
torch
torchvision
mlflow
boto3
python-multipart
